NXP-22553: Fix Mime type handling in webengine
