+++ nuxeo-webengine/nuxeo-webengine-core/src/main/java/org/nuxeo/ecm/webengine/app/jersey/WebEngineServlet.java
@@ -43,104 +43,6 @@ import org.nuxeo.ecm.webengine.jaxrs.servlet.ApplicationServlet;
  */
 public class WebEngineServlet extends ApplicationServlet {
 
-    protected final String MIME_TYPE = "X-File-Type";
-
-    private final class DefaultContentTypeRequestWrapper extends HttpServletRequestWrapper {
-
-        protected final Hashtable<String, String[]> headers;
-
-        protected final String lCONTENT_TYPE = HttpHeaders.CONTENT_TYPE.toLowerCase();
-
-        protected final String lFILE_TYPE = MIME_TYPE.toLowerCase();
-
-        protected DefaultContentTypeRequestWrapper(HttpServletRequest request, boolean patchCType, boolean patchMType) {
-            super(request);
-            headers = patchHeaders(request, patchCType, patchMType);
-        }
-
-        protected Hashtable<String, String[]> patchHeaders(HttpServletRequest request, boolean patchCType,
-                boolean patchMType) {
-            Hashtable<String, String[]> headers = new Hashtable<>();
-            // collect headers from request
-            Enumeration<String> eachNames = request.getHeaderNames();
-            while (eachNames.hasMoreElements()) {
-                String name = eachNames.nextElement().toLowerCase();
-                List<String> values = new LinkedList<>();
-                Enumeration<String> eachValues = request.getHeaders(name);
-                while (eachValues.hasMoreElements()) {
-                    values.add(eachValues.nextElement());
-                }
-                headers.put(name, values.toArray(new String[values.size()]));
-            }
-            if (patchCType) {
-                // patch content type
-                String ctype = request.getContentType();
-                if (ctype == null || ctype.isEmpty()) {
-                    String[] ctypes = new String[] { "application/octet-stream" };
-                    headers.put(lCONTENT_TYPE, ctypes);
-                } else {
-                    patchContentTypes(headers.get(lCONTENT_TYPE));
-                }
-            }
-            if (patchMType) {
-                patchContentTypes(headers.get(lFILE_TYPE));
-            }
-            return headers;
-        }
-
-        protected void patchContentTypes(String[] ctypes) {
-            for (int index = 0; index < ctypes.length; ++index) {
-                String value = ctypes[index];
-                if (value.isEmpty()) {
-                    ctypes[index] = "application/octet-stream";
-                } else if (!value.contains("/")) {
-                    ctypes[index] = "application/".concat(value);
-                }
-            }
-        }
-
-        @Override
-        public Enumeration<String> getHeaderNames() {
-            return headers.keys();
-        }
-
-        @Override
-        public String getHeader(String name) {
-            String lname = name.toLowerCase();
-            if (!headers.containsKey(lname)) {
-                return null;
-            }
-            return headers.get(lname)[0];
-        }
-
-        @Override
-        public Enumeration<String> getHeaders(final String name) {
-            final String lname = name.toLowerCase();
-            if (!headers.containsKey(lname)) {
-                return Collections.emptyEnumeration();
-            }
-            return new Enumeration<String>() {
-                String[] values = headers.get(lname);
-
-                int index = 0;
-
-                @Override
-                public boolean hasMoreElements() {
-                    return index < values.length;
-                }
-
-                @Override
-                public String nextElement() {
-                    if (index >= values.length) {
-                        throw new NoSuchElementException(index + " is higher than " + values.length);
-                    }
-                    return values[index++];
-                }
-            };
-        }
-
-    }
-
     private static final long serialVersionUID = 1L;
 
     @Override
@@ -166,13 +68,6 @@ public class WebEngineServlet extends ApplicationServlet {
             // character
             // from the input stream - see WebComponent.isEntityPresent.
             request.getParameterMap();
-        }
-        final String ctype = request.getHeader(HttpHeaders.CONTENT_TYPE);
-        final String mtype = request.getHeader(MIME_TYPE);
-        boolean patchCType = ctype == null || ctype.length() == 0 || !ctype.contains("/");
-        boolean patchMMType = mtype != null && !mtype.contains("/");
-        if (patchCType || patchMMType) {
-            request = new DefaultContentTypeRequestWrapper(request, patchCType, patchMMType);
         }
         container.service(request, response);
     }
