+++ nuxeo-webengine/nuxeo-webengine-core/src/main/resources/OSGI-INF/deployment-fragment.xml
@@ -56,6 +56,20 @@
     </servlet-mapping>
 
     <filter>
+      <display-name>HeaderFixFilter</display-name>
+      <filter-name>HeaderFixFilter</filter-name>
+      <filter-class>
+        org.nuxeo.ecm.webengine.app.HeaderFixFilter
+      </filter-class>
+    </filter>
+    <filter-mapping>
+      <filter-name>HeaderFixFilter</filter-name>
+      <url-pattern>/site/*</url-pattern>
+      <dispatcher>REQUEST</dispatcher>
+      <dispatcher>FORWARD</dispatcher>
+    </filter-mapping>
+
+    <filter>
       <display-name>RequestContextFilter</display-name>
       <filter-name>RequestContextFilter</filter-name>
       <filter-class>
